+++ site/e2e/global.setup.ts
@@ -1,7 +1,7 @@
 import { test, expect } from "@playwright/test"
 import * as constants from "./constants"
 import { STORAGE_STATE } from "./playwright.config"
-import { Language } from "../src/components/CreateUserForm/CreateUserForm"
+import { Language } from "../src/pages/CreateUserPage/CreateUserForm/CreateUserForm"
 
 test("create first user", async ({ page }) => {
   await page.goto("/", { waitUntil: "domcontentloaded" })
