+++ site/src/pages/CreateUserPage/CreateUserPage.tsx
@@ -4,10 +4,10 @@ import { FC } from "react"
 import { Helmet } from "react-helmet-async"
 import { useNavigate } from "react-router-dom"
 import { createUserMachine } from "xServices/users/createUserXService"
-import * as TypesGen from "../../../api/typesGenerated"
-import { CreateUserForm } from "../../../components/CreateUserForm/CreateUserForm"
-import { Margins } from "../../../components/Margins/Margins"
-import { pageTitle } from "../../../utils/page"
+import * as TypesGen from "../../api/typesGenerated"
+import { CreateUserForm } from "./CreateUserForm/CreateUserForm"
+import { Margins } from "../../components/Margins/Margins"
+import { pageTitle } from "../../utils/page"
 import { getAuthMethods } from "api/api"
 import { useQuery } from "@tanstack/react-query"
